# server/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Event Management API - Frontend Consumed Endpoints",
    "version": "1.0.0",
    "description": "Comprehensive API documentation for backend endpoints consumed by Homepage, Category pages, Payment page, User Dashboard, and Contact Us page.",
    "contact": {
      "name": "Event Management API Support",
      "email": "shriansh.j23@iiits.in"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie set after successful authentication. Required for protected user/payment endpoints."
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "integer",
                "example": 400
              },
              "message": {
                "type": "string",
                "example": "Invalid event ID format"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
      "EventListItem": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "67fa7a1f0c551c9170b9bfe1"
          },
          "category": {
            "type": "string",
            "example": "TEDx"
          },
          "title": {
            "type": "string",
            "example": "TEDx Future of AI"
          },
          "description": {
            "type": "string",
            "example": "Ideas worth spreading on AI and ethics."
          },
          "startDateTime": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-10T09:00:00.000Z"
          },
          "endDateTime": {
            "type": "string",
            "format": "date-time",
            "example": "2026-04-10T16:00:00.000Z"
          },
          "venue": {
            "type": "string",
            "example": "Convention Center Hall A"
          },
          "capacity": {
            "type": "integer",
            "example": 500
          },
          "ticketPrice": {
            "type": "number",
            "example": 49.99
          },
          "status": {
            "type": "string",
            "enum": [
              "start_selling",
              "upcoming",
              "over"
            ],
            "example": "start_selling"
          },
          "organizerId": {
            "type": "string",
            "example": "67fa7a1f0c551c9170b9bf11"
          },
          "image": {
            "type": "string",
            "nullable": true,
            "example": "uploads/events/tedx-banner.jpg"
          }
        },
        "required": [
          "_id",
          "category",
          "title",
          "startDateTime",
          "venue",
          "ticketPrice",
          "status"
        ]
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "currentPage": {
            "type": "integer",
            "example": 1
          },
          "totalPages": {
            "type": "integer",
            "example": 4
          },
          "totalEvents": {
            "type": "integer",
            "example": 12
          },
          "hasNextPage": {
            "type": "boolean",
            "example": true
          },
          "hasPrevPage": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "currentPage",
          "totalPages",
          "totalEvents",
          "hasNextPage",
          "hasPrevPage"
        ]
      },
      "UserDashboardBooking": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "example": "67fa7a1f0c551c9170b9bfe1"
          },
          "title": {
            "type": "string",
            "example": "TEDx Future of AI"
          },
          "startDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "endDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "venue": {
            "type": "string",
            "example": "Convention Center Hall A"
          },
          "ticketType": {
            "type": "string",
            "example": "Standard"
          },
          "price": {
            "type": "number",
            "example": 49.99
          },
          "status": {
            "type": "string",
            "example": "upcoming"
          },
          "registrationStatus": {
            "type": "string",
            "enum": [
              "active",
              "cancelled"
            ],
            "example": "active"
          },
          "ticketCount": {
            "type": "integer",
            "example": 2
          },
          "activeTicketCount": {
            "type": "integer",
            "example": 2
          },
          "cancelledTicketCount": {
            "type": "integer",
            "example": 0
          },
          "registrationIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "67fa7b271688151a7cd61d2a",
              "67fa7b271688151a7cd61d2b"
            ]
          },
          "activeRegistrationIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "67fa7b271688151a7cd61d2a",
              "67fa7b271688151a7cd61d2b"
            ]
          },
          "registrationDate": {
            "type": "string",
            "format": "date-time"
          },
          "canCancel": {
            "type": "boolean",
            "example": true
          },
          "refundPercentage": {
            "type": "number",
            "example": 50
          },
          "refundAmountPerTicket": {
            "type": "number",
            "example": 24.995
          },
          "totalRefundedAmount": {
            "type": "number",
            "example": 0
          }
        }
      },
...
    }
  },
  "paths": {
    "/events": {
      "get": {
        "tags": [
          "Homepage"
        ],
        "summary": "Fetch events for Home page sections",
        "description": "Returns events grouped into two streams consumed by Home page: `selling` (`start_selling`) and `upcoming` (`upcoming`) with independent pagination.",
        "parameters": [
          {
            "in": "query",
            "name": "sellingPage",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for `selling` section."
          },
          {
            "in": "query",
            "name": "upcomingPage",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for `upcoming` section."
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 12
            },
            "description": "Items per section per page."
          },
          {
            "in": "query",
            "name": "title",
            "schema": {
              "type": "string"
            },
            "description": "Case-insensitive title filter."
          },
          {
            "in": "query",
            "name": "venue",
            "schema": {
              "type": "string"
            },
            "description": "Case-insensitive venue filter."
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "Case-insensitive category filter."
          }
        ],
        "responses": {
          "200": {
            "description": "Home page event data retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "selling": {
                          "type": "object",
                          "properties": {
                            "events": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/EventListItem"
                              }
                            },
                            "pagination": {
                              "$ref": "#/components/schemas/PaginationMeta"
                            }
                          }
                        },
                        "upcoming": {
                          "type": "object",
                          "properties": {
                            "events": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/EventListItem"
                              }
                            },
                            "pagination": {
                              "$ref": "#/components/schemas/PaginationMeta"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error while fetching events.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/stats": {
      "get": {
        "tags": [
          "Homepage"
        ],
        "summary": "Get hero statistics",
        "description": "Returns aggregate counts displayed in homepage hero (events and organizers).",
        "responses": {
          "200": {
            "description": "Stats fetched successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "totalEvents": {
                          "type": "integer",
                          "example": 124
                        },
                        "totalOrganizers": {
                          "type": "integer",
                          "example": 42
                        }
                      },
                      "required": [
                        "totalEvents",
                        "totalOrganizers"
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch stats.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "An error occurred while fetching statistics."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/user/dashboard": {
      "get": {
        "tags": [
          "User Dashboard"
        ],
        "summary": "Get consolidated dashboard payload",
        "description": "Returns user profile basics, grouped bookings, and dashboard stats for overview and booking tabs.",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard data returned.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "name": {
                              "type": "string",
                              "example": "Aarav Sharma"
                            },
                            "email": {
                              "type": "string",
                              "format": "email",
                              "example": "aarav@example.com"
                            },
                            "username": {
                              "type": "string",
                              "example": "aaravsharma"
                            }
                          }
                        },
                        "bookings": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/UserDashboardBooking"
                          }
                        },
                        "stats": {
                          "type": "object",
                          "properties": {
                            "totalBookings": {
                              "type": "integer",
                              "example": 5
                            },
                            "upcomingEvents": {
                              "type": "integer",
                              "example": 3
                            },
                            "completedEvents": {
                              "type": "integer",
                              "example": 1
                            },
                            "cancelledEvents": {
                              "type": "integer",
                              "example": 1
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "User not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "User not authenticated"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/contact": {
      "post": {
        "tags": [
          "Contact Us"
        ],
        "summary": "Submit contact form",
        "description": "Accepts contact form payload from frontend Contact Us page and attempts to notify support email (simulated when email config is absent).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "subject",
                  "message"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 100,
                    "example": "Riya Verma"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "riya@example.com"
                  },
                  "subject": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 150,
                    "example": "Question about event refunds"
                  },
                  "message": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 3000,
                    "example": "I wanted to understand refund eligibility if I cancel two days before the event."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact message accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Your message has been received. Our team will contact you shortly."
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "submittedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "ticketId": {
                          "type": "string",
                          "example": "CT-1711801000000-ABCD12"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "name, email, subject and message are required"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    }
  }
}